Extract collections logic to selector
