+++ packages/mobile/src/screens/edit-track-screen/components/CollectibleGatedAvailability.tsx
@@ -1,4 +1,4 @@
-import { useCallback, useEffect, useMemo } from 'react'
+import { useCallback, useEffect } from 'react'
 
 import type { Nullable, PremiumConditions } from '@audius/common'
 import { useField } from 'formik'
